Create nav.html
Put the navigation bar in the static directory to simplify code. Now, any time we want to change the nav bar, we just have to change static/nav.html and it will update all nav bars that use it. 

Note: any change to a file in the static directory will require clearing the cache of the browser to update it.