Create footer.html
Put footer.html in the static directory to simplify code. Now, any time we want to change the footer, we just have to change static/footer.html and it will update all footers that use it. 

Note: any change to a file in the static directory will require clearing the cache of the browser to update it.